Move naf.py to provider module.
